requirements.txt
flask
pymongo
paho-mqtt
pyyaml
cryptography
